fix: restore DateFormatter.theMovieDatabase for encoder timezone consistency
The JSONEncoder.theMovieDatabase was changed to use Date.ISO8601FormatStyle
which outputs dates in UTC, while the decoder uses .autoupdatingCurrent
(local timezone). This caused a 1-day offset in encode/decode round-trips
for users in non-UTC timezones (e.g. BST UTC+1).

Restore DateFormatter.theMovieDatabase in sources so the encoder and decoder
use the same local timezone, preserving existing behaviour. Also removes the
now-unnecessary test-only copy of the formatter.
