+++ Sources/TMDb/Extensions/DateFormatter+TMDb.swift
@@ -9,6 +9,13 @@ import Foundation
 
 extension DateFormatter {
 
+    static var theMovieDatabase: DateFormatter {
+        let dateFormatter = DateFormatter()
+        dateFormatter.locale = Locale(identifier: "en_US_POSIX")
+        dateFormatter.dateFormat = "yyyy-MM-dd"
+        return dateFormatter
+    }
+
     static var theMovieDatabaseAuth: DateFormatter {
         let dateFormatter = DateFormatter()
         dateFormatter.locale = Locale(identifier: "en_US_POSIX")
